requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
httplib2==0.22.0
idna==3.8
oauthlib==3.2.2
outcome==1.3.0.post0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
PyPDF2==3.0.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
